% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{BMI}
\usage{
bmi(weight, height, name)
}
\arguments{
\item{weight}{number}

\item{height}{number}

\item{name}{your name}
}
\value{
bmi
}
\description{
Lorem ipsum dolor sit meat, et mei falli pungent, e
am at audire vivendo. No tollit postulant dissentiunt quo,
mea tractatos salutandi u
}
\examples{
library(health)
bmi(70, 173, "John")

}
