% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health-package.R
\docType{package}
\name{health-package}
\alias{health}
\alias{health-package}
\title{health: Daniel Vartanian Health Data Package}
\description{
This package contains all my (Daniel Vartanian) health data. It 
    represents a personal project that aims to monitor my body's expression 
    over time and create a free public record for posterity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/health/}
  \item \url{https://github.com/danielvartan/health/}
  \item Report bugs at \url{https://github.com/danielvartan/health/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

}
\keyword{internal}
