% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\docType{data}
\name{people}
\alias{people}
\title{people data}
\format{
A data frame with 27 rows and 3 variables

\itemize{
  \item weight. weight(kg)
  \item height. height(cm)
  \item age. age
}
}
\source{
LIN WEI HAN,2023
}
\usage{
data(people)
}
\description{
A dataset description
}
\references{
LIN WEI HAN,2023
}
